server/user_service/requirements.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-environ==0.12.0
django_csp==3.8
djangorestframework==3.15.2
kombu==5.5.1
prompt_toolkit==3.0.50
psycopg==3.2.6
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
